+++ src/main/java/kr/co/broadwave/aci/equipment/EquipmentMapperDto.java
@@ -27,7 +27,7 @@ public class EquipmentMapperDto {
     private String emAwsNumber; // AWS상 Device ID
     private String emEmbeddedNumber; // 임베디드 기판 번호
     private String operator;
-    private Long emMaximumPayload;
+    private Double emMaximumPayload;
     private Long emUnit;
 
     public String getEmNumber() {
@@ -66,7 +66,7 @@ public class EquipmentMapperDto {
         return operator;
     }
 
-    public Long getEmMaximumPayload() {
+    public Double getEmMaximumPayload() {
         return emMaximumPayload;
     }
